requirements.txt
requests>=2.31.0
pandas>=2.2.0
google-api-python-client>=2.125.0
google-auth>=2.29.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0
openai>=1.30.0
anthropic>=0.25.0
python-dotenv>=1.0.0
apify-client>=1.6.0
colorama>=0.4.6
tqdm>=4.66.0